Separate between unit and integration tests.
Also, fix some issues that existed with the tests themselves. For unit
tests now we have the Makefile target `test`, and for integration tests
`integration`.
