+++ pkg/apis/kubic/v1beta1/v1beta1_suite_test.go
@@ -33,6 +33,10 @@ var cfg *rest.Config
 var c client.Client
 
 func TestMain(m *testing.M) {
+	if testing.Short() {
+		return
+	}
+
 	t := &envtest.Environment{
 		CRDDirectoryPaths: []string{filepath.Join("..", "..", "..", "..", "configs", "crds")},
 	}
